# perf: arm: don't init `tmp_buf` before calling padding and filter funcs