Hi,

Took a look at the `rav1d` vs `dav1d` perf contest, and was able to find a small (~1%) improvement (as measured on my M3 MacBook Air).

This PR includes two changes:
1. In the `cdef_filter_neon_erased` call, replace `Align16([0; TMP_LEN])` with `Align16([MaybeUninit::uninit(); TMP_LEN])`.
2. In `rav1d_cdef_brow`, move up the `let mut lr_bak = Align16([..])` initialization to be outside the loop. (now in #1399 )

I believe these should be safe and correct since the C version doesn't seem to initialize `ALIGN_STK_16`-created data, but this should definitely be double-checked by someone more familiar with the code 🙏 

These changes are the result of comparing runs of `dav1d` and `rav1d`, which showed:

1. `cdef_filter_neon_erased` had a much larger "self" time vs. `cdef_filter_8x8_neon`+`cdef_filter_4x4_neon`, which seems to be due to a `memset`.
3. Similarly, `rav1d::src::cdef_apply::rav1d_cdef_brow` has a slighly larger self time compared to `dav1d_cdef_brow_8bpc`.

### Local hyperfine output

`main` (commit a654c1e82adb2d9a33ae50d2a82a7a747102cbb6):
```
$ hyperfine --warmup 2 "target/release/dav1d -q -i Chimera-AV1-8bit-1920x1080-6736kbps.ivf -o /dev/null --threads 1"
Benchmark 1: target/release/dav1d -q -i Chimera-AV1-8bit-1920x1080-6736kbps.ivf -o /dev/null --threads 1
  Time (mean ± σ):     73.914 s ±  0.151 s    [User: 73.295 s, System: 0.279 s]
  Range (min … max):   73.770 s … 74.132 s    10 runs
```

this branch:
```
$ hyperfine --warmup 2 "target/release/dav1d -q -i Chimera-AV1-8bit-1920x1080-6736kbps.ivf -o /dev/null --threads 1"
Benchmark 1: target/release/dav1d -q -i Chimera-AV1-8bit-1920x1080-6736kbps.ivf -o /dev/null --threads 1
  Time (mean ± σ):     72.892 s ±  0.103 s    [User: 72.314 s, System: 0.269 s]
  Range (min … max):   72.736 s … 73.048 s    10 runs
```