@ohadravid 2025-05-19
@kkysen changed the code to always pass around `MaybeUninit` (slice/raw ptr) to avoid the unsoundness, as a separate commit for now, let me know how/if you prefer a different approach 🐛