@kkysen 2025-05-20
> @kkysen changed the code to always pass around `MaybeUninit` (slice/raw ptr) to avoid the unsoundness, as a separate commit for now, let me know how/if you prefer a different approach 🐛

Can you squash those two commits together so we don't have an unsound one?  Thanks!

After adding the docs, too, this should be good to go.